- move server out of main
